requirements.txt
fastapi
httpx
numpy
openai
pandas
pytest
python-dotenv
requests
uvicorn
sqlalchemy
email-validator
passlib
bcrypt=3.1.7
python-jose